DROID-3930 refactoring
